Adding Day 11 - Blackjack cornerstone
Mine (blackjack.py) isn't as clean as Angela's but I'm happy with how I've achieved the solution.

There is certainly some simplification that I can take away from Angela's solution in particular around how to deal the first set of cards - a for loop is a much better option than the while loop. Also comparing and deciding the winner is much cleaner in Angela's solution.